rename util module
