+++ tests/unit/runtime/compile/test_wrapper.py
@@ -9,7 +9,7 @@ import torch
 import deepspeed
 from deepspeed.accelerator import get_accelerator
 
-from utils import DistributedCompileTest
+from tests.unit.runtime.compile.common import DistributedCompileTest
 
 
 @pytest.fixture
